up category, phan trang cho shop...

+++ src/main/java/org/example/casestudymodule4shoestore/repositories/products/IProductRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
+import java.util.Map;
 
 public interface IProductRepository extends JpaRepository<Product, Long> {
 
@@ -16,7 +17,7 @@ public interface IProductRepository extends JpaRepository<Product, Long> {
 //    @Query( nativeQuery = true, value = "select p.id as id, p.name as name, p.image as image, p.price as price, c.id as categoryId, c.name as categoryName from products p join categories c on p.cat_id = c.id where c.id = :id")
 //    List<ProductCategoryDTO> findByCategory(@Param("id") Long id);
 
-    List<Product> findAllByCat_Id(Integer id);
+    List<Product> findAllByCat_Id(Long id);
 
     @Query( nativeQuery = true, value = "select p.*  from products as p where p.name like :keyword")
     List<Product> findAllByNameContaining(@Param("keyword")String keyword);
@@ -48,5 +49,6 @@ public interface IProductRepository extends JpaRepository<Product, Long> {
 
     @Query(nativeQuery = true, value = "SELECT cart_id FROM cart WHERE customer_id = :idCustomer")
     int findIdCart(@Param("idCustomer") int idCustomer);
+
 }
